+++ source/xzs/src/main/java/com/mindskip/xzs/controller/wx/student/ExamPaperAnswerController.java
@@ -26,6 +26,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -76,7 +77,7 @@ public class ExamPaperAnswerController extends BaseWXApiController {
             return RestResponse.fail(2, "试卷不能重复做");
         }
         ExamPaperAnswer examPaperAnswer = examPaperAnswerInfo.getExamPaperAnswer();
-        Integer userScore = examPaperAnswer.getUserScore();
+        BigDecimal userScore = examPaperAnswer.getUserScore();
         String scoreVm = ExamUtil.scoreToVM(userScore);
         UserEventLog userEventLog = new UserEventLog(user.getId(), user.getUserName(), user.getRealName(), new Date());
         String content = user.getUserName() + " 提交试卷：" + examPaperAnswerInfo.getExamPaper().getName()
