+++ source/xzs/src/main/java/com/mindskip/xzs/domain/other/ExamPaperAnswerUpdate.java
@@ -4,11 +4,13 @@ package com.mindskip.xzs.domain.other;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class ExamPaperAnswerUpdate {
     private Integer id;
-    private Integer customerScore;
+    private BigDecimal customerScore;
     private Boolean doRight;
 
 }
